web/requirements.txt
# requirements.txt
Flask==1.0.2
uWSGI==2.0.17.1
pymongo >= 2.1
flask-cors
redis
python-dotenv
sentry-sdk[flask]==0.7.14
flask-restplus
mongoengine